requirements.txt
azure-functions
Pillow
tensorflow
h5py
